uint8

+++ silo-core/contracts/Silo.sol
@@ -101,8 +101,8 @@ contract Silo is SiloERC4626 {
     /// @inheritdoc ISilo
     function utilizationData() external view virtual returns (UtilizationData memory) {
         return UtilizationData({
-            collateralAssets: total[AssetType.Collateral].assets,
-            debtAssets: total[AssetType.Debt].assets,
+            collateralAssets: total[uint8(AssetType.Collateral)].assets,
+            debtAssets: total[uint8(AssetType.Debt)].assets,
             interestRateTimestamp: siloData.interestRateTimestamp
         });
     }
@@ -150,8 +150,8 @@ contract Silo is SiloERC4626 {
         virtual
         returns (uint256 totalCollateralAssets, uint256 totalProtectedAssets)
     {
-        totalCollateralAssets = total[AssetType.Collateral].assets;
-        totalProtectedAssets = total[AssetType.Protected].assets;
+        totalCollateralAssets = total[uint8(AssetType.Collateral)].assets;
+        totalProtectedAssets = total[uint8(AssetType.Protected)].assets;
     }
 
     /// @inheritdoc ISilo
@@ -161,8 +161,8 @@ contract Silo is SiloERC4626 {
         virtual
         returns (uint256 totalCollateralAssets, uint256 totalDebtAssets)
     {
-        totalCollateralAssets = total[AssetType.Collateral].assets;
-        totalDebtAssets = total[AssetType.Debt].assets;
+        totalCollateralAssets = total[uint8(AssetType.Collateral)].assets;
+        totalDebtAssets = total[uint8(AssetType.Debt)].assets;
     }
 
     // ERC4626
@@ -203,7 +203,7 @@ contract Silo is SiloERC4626 {
 
     /// @inheritdoc IERC4626
     function maxDeposit(address /* _receiver */) external view virtual returns (uint256 maxAssets) {
-        return _callMaxDepositOrMint(total[AssetType.Collateral].assets);
+        return _callMaxDepositOrMint(total[uint8(AssetType.Collateral)].assets);
     }
 
     /// @inheritdoc IERC4626
@@ -305,7 +305,7 @@ contract Silo is SiloERC4626 {
         virtual
         returns (uint256 maxAssets)
     {
-        return _callMaxDepositOrMint(total[_assetType].assets);
+        return _callMaxDepositOrMint(total[uint8(_assetType)].assets);
     }
 
     /// @inheritdoc ISilo
@@ -454,9 +454,9 @@ contract Silo is SiloERC4626 {
             _borrowAssets,
             _borrower,
             _assetType,
-            total[AssetType.Collateral].assets,
-            total[AssetType.Debt],
-            total[_assetType]
+            total[uint8(AssetType.Collateral)].assets,
+            total[uint8(AssetType.Debt)],
+            total[uint8(_assetType)]
         );
 
         emit Borrow(msg.sender, _borrower, _borrower, _borrowAssets, borrowedShares);
@@ -623,7 +623,7 @@ contract Silo is SiloERC4626 {
 
     // TODO can we optimise this? maybe add as args to methods
     function getRawLiquidity() public view virtual returns (uint256 liquidity) {
-        liquidity = SiloMathLib.liquidity(total[AssetType.Collateral].assets, total[AssetType.Debt].assets);
+        liquidity = SiloMathLib.liquidity(total[uint8(AssetType.Collateral)].assets, total[uint8(AssetType.Debt)].assets);
     }
 
     /// @dev that method allow to finish liquidation process by giving up collateral to liquidator
@@ -657,7 +657,7 @@ contract Silo is SiloERC4626 {
         returns (uint256 assets, uint256 shares)
     {
         // TODO make sure total are updated when we call accrue interest from SiloConfig
-        (assets, shares) = Actions.deposit( sharedStorage, _assets, _shares, _receiver, _assetType, total[_assetType]);
+        (assets, shares) = Actions.deposit( sharedStorage, _assets, _shares, _receiver, _assetType, total[uint8(_assetType)]);
 
         if (_assetType == AssetType.Collateral) {
             emit Deposit(msg.sender, _receiver, assets, shares);
@@ -689,7 +689,7 @@ contract Silo is SiloERC4626 {
                 spender: _spender,
                 assetType: _assetType
             }),
-            total[_assetType]
+            total[uint8(_assetType)]
         );
 
         if (_assetType == AssetType.Collateral) {
@@ -721,9 +721,9 @@ contract Silo is SiloERC4626 {
                 borrower: _borrower,
                 sameAsset: _sameAsset,
                 leverage: _leverage,
-                totalCollateralAssets: total[AssetType.Collateral].assets
+                totalCollateralAssets: total[uint8(AssetType.Collateral)].assets
             }),
-            total[AssetType.Debt],
+            total[uint8(AssetType.Debt)],
             _data
         );
 
@@ -745,7 +745,7 @@ contract Silo is SiloERC4626 {
             _borrower,
             _repayer,
             _liquidation,
-            total[AssetType.Debt]
+            total[uint8(AssetType.Debt)]
         );
 
         emit Repay(_repayer, _borrower, assets, shares);
@@ -829,7 +829,7 @@ contract Silo is SiloERC4626 {
             _config,
             _owner,
             _assetType,
-            _assetType == AssetType.Protected ? total[AssetType.Protected].assets : 0 // will be calculated internally
+            _assetType == AssetType.Protected ? total[uint8(AssetType.Protected)].assets : 0 // will be calculated internally
         );
     }
 
@@ -848,8 +848,8 @@ contract Silo is SiloERC4626 {
             _daoFee,
             _deployerFee,
             siloData,
-            total[AssetType.Collateral],
-            total[AssetType.Debt]
+            total[uint8(AssetType.Collateral)],
+            total[uint8(AssetType.Debt)]
         );
 
         if (accruedInterest != 0) emit AccruedInterest(accruedInterest);
